+++ app/Http/Controllers/OptionsController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\View\View;
+
+class OptionsController
+{
+    public function index(): View
+    {
+        return view('Options');
+    }
+}