+++ resources/views/components/site-footer.blade.php
@@ -1,0 +1,3 @@
+<footer class="site-footer">
+    <p>YouTube RSS</p>
+</footer>