+++ routes/web.php
@@ -1,11 +1,17 @@
 <?php
 
+use App\Http\Controllers\OptionsController;
 use App\Http\Controllers\YoutubeChannelController;
 use Illuminate\Support\Facades\Route;
 
-//Main Page
+// Channels page
 Route::get('/', [YoutubeChannelController::class, 'index'])
     ->name('index');
 
+// Feeds
 Route::get('/feeds/{youtubeChannel:youtube_id}.xml', [YoutubeChannelController::class, 'feed'])
     ->name('feeds.show');
+
+// Options
+Route::get('/options', [OptionsController::class, 'index'])
+    ->name('options');